Adding full API class
